feat: refactoring for ctype tests (currently broken)
