+++ pallets/ctype/src/benchmarking.rs
@@ -1,51 +1,51 @@
-// KILT Blockchain – https://botlabs.org
-// Copyright (C) 2019-2021 BOTLabs GmbH
+// // KILT Blockchain – https://botlabs.org
+// // Copyright (C) 2019-2021 BOTLabs GmbH
 
-// The KILT Blockchain is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
+// // The KILT Blockchain is free software: you can redistribute it and/or modify
+// // it under the terms of the GNU General Public License as published by
+// // the Free Software Foundation, either version 3 of the License, or
+// // (at your option) any later version.
 
-// The KILT Blockchain is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
+// // The KILT Blockchain is distributed in the hope that it will be useful,
+// // but WITHOUT ANY WARRANTY; without even the implied warranty of
+// // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// // GNU General Public License for more details.
 
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+// // You should have received a copy of the GNU General Public License
+// // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-// If you feel like getting in touch with us, you can do so at info@botlabs.org
+// // If you feel like getting in touch with us, you can do so at info@botlabs.org
 
-#![cfg(feature = "runtime-benchmarks")]
+// #![cfg(feature = "runtime-benchmarks")]
 
-use super::*;
+// use super::*;
 
-use frame_benchmarking::{account, benchmarks};
-use frame_system::RawOrigin;
+// use frame_benchmarking::{account, benchmarks};
+// use frame_system::RawOrigin;
 
-const SEED: u32 = 0;
+// const SEED: u32 = 0;
 
-benchmarks! {
-	add {
-		let caller = account("caller", 0, SEED);
-		let hash = <T::Hash as Default>::default();
+// benchmarks! {
+// 	add {
+// 		let caller = account("caller", 0, SEED);
+// 		let hash = <T::Hash as Default>::default();
 
-	}: _(RawOrigin::Signed(caller), hash)
-	verify {
-		CTYPEs::<T>::contains_key(hash)
-	}
-}
+// 	}: _(RawOrigin::Signed(caller), hash)
+// 	verify {
+// 		CTYPEs::<T>::contains_key(hash)
+// 	}
+// }
 
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use crate::tests::{new_test_ext, Test};
-	use frame_support::assert_ok;
+// #[cfg(test)]
+// mod tests {
+// 	use super::*;
+// 	use crate::tests::{new_test_ext, Test};
+// 	use frame_support::assert_ok;
 
-	#[test]
-	fn test_benchmarks() {
-		new_test_ext().execute_with(|| {
-			assert_ok!(test_benchmark_add::<Test>());
-		});
-	}
-}
+// 	#[test]
+// 	fn test_benchmarks() {
+// 		new_test_ext().execute_with(|| {
+// 			assert_ok!(test_benchmark_add::<Test>());
+// 		});
+// 	}
+// }
