+++ pallets/ctype/src/tests.rs
@@ -16,134 +16,25 @@
 
 // If you feel like getting in touch with us, you can do so at info@botlabs.org
 
-use crate as pallet_ctype;
-use crate::*;
-
-use frame_support::{
-	assert_noop, assert_ok, parameter_types,
-	weights::{
-		constants::{BlockExecutionWeight, ExtrinsicBaseWeight, RocksDbWeight, WEIGHT_PER_SECOND},
-		DispatchClass, Weight,
-	},
-};
-use frame_system::limits::{BlockLength, BlockWeights};
-use kilt_primitives::Signature;
-use sp_core::{ed25519, Pair, H256};
-use sp_runtime::{
-	testing::Header,
-	traits::{BlakeTwo256, IdentifyAccount, IdentityLookup, Saturating, Verify},
-	MultiSigner, Perbill,
-};
-
-type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
-type Block = frame_system::mocking::MockBlock<Test>;
-
-frame_support::construct_runtime!(
-	pub enum Test where
-		Block = Block,
-		NodeBlock = Block,
-		UncheckedExtrinsic = UncheckedExtrinsic,
-	{
-		System: frame_system::{Pallet, Call, Config, Storage, Event<T>},
-		Ctype: pallet_ctype::{Pallet, Call, Storage, Event<T>},
-	}
-);
-
-parameter_types! {
-	pub const BlockHashCount: u64 = 250;
-	pub const MaximumBlockWeight: Weight = 2 * WEIGHT_PER_SECOND;
-	pub const MaximumBlockLength: u32 = 5 * 1024 * 1024;
-	pub const AvailableBlockRatio: Perbill = Perbill::from_percent(75);
-	pub MaximumExtrinsicWeight: Weight = AvailableBlockRatio::get()
-		.saturating_sub(Perbill::from_percent(10)) * MaximumBlockWeight::get();
-}
-
-/// We assume that ~10% of the block weight is consumed by `on_initalize`
-/// handlers. This is used to limit the maximal weight of a single extrinsic.
-const AVERAGE_ON_INITIALIZE_RATIO: Perbill = Perbill::from_percent(10);
-/// We allow `Normal` extrinsics to fill up the block up to 75%, the rest can be
-/// used by  Operational  extrinsics.
-const NORMAL_DISPATCH_RATIO: Perbill = Perbill::from_percent(75);
-/// We allow for 2 seconds of compute with a 6 second average block time.
-const MAXIMUM_BLOCK_WEIGHT: Weight = 2 * WEIGHT_PER_SECOND;
-
-parameter_types! {
-	pub RuntimeBlockLength: BlockLength =
-		BlockLength::max_with_normal_ratio(5 * 1024 * 1024, NORMAL_DISPATCH_RATIO);
-	pub RuntimeBlockWeights: BlockWeights = BlockWeights::builder()
-		.base_block(BlockExecutionWeight::get())
-		.for_class(DispatchClass::all(), |weights| {
-			weights.base_extrinsic = ExtrinsicBaseWeight::get();
-		})
-		.for_class(DispatchClass::Normal, |weights| {
-			weights.max_total = Some(NORMAL_DISPATCH_RATIO * MAXIMUM_BLOCK_WEIGHT);
-		})
-		.for_class(DispatchClass::Operational, |weights| {
-			weights.max_total = Some(MAXIMUM_BLOCK_WEIGHT);
-			// Operational transactions have some extra reserved space, so that they
-			// are included even if block reached `MAXIMUM_BLOCK_WEIGHT`.
-			weights.reserved = Some(
-				MAXIMUM_BLOCK_WEIGHT - NORMAL_DISPATCH_RATIO * MAXIMUM_BLOCK_WEIGHT
-			);
-		})
-		.avg_block_initialization(AVERAGE_ON_INITIALIZE_RATIO)
-		.build_or_panic();
-	pub const SS58Prefix: u8 = 38;
-}
-
-impl frame_system::Config for Test {
-	type Origin = Origin;
-	type Call = Call;
-	type Index = u64;
-	type BlockNumber = u64;
-	type Hash = H256;
-	type Hashing = BlakeTwo256;
-	type AccountId = <<Signature as Verify>::Signer as IdentifyAccount>::AccountId;
-	type Lookup = IdentityLookup<Self::AccountId>;
-	type Header = Header;
-	type Event = ();
-	type BlockHashCount = BlockHashCount;
-	type DbWeight = RocksDbWeight;
-	type Version = ();
-
-	type PalletInfo = PalletInfo;
-	type AccountData = ();
-	type OnNewAccount = ();
-	type OnKilledAccount = ();
-	type BaseCallFilter = ();
-	type SystemWeightInfo = ();
-	type BlockWeights = RuntimeBlockWeights;
-	type BlockLength = RuntimeBlockLength;
-	type SS58Prefix = SS58Prefix;
-	type OnSetCode = ();
-}
-
-impl Config for Test {
-	type Event = ();
-	type WeightInfo = ();
-}
-
-type CType = Module<Test>;
-
-pub fn new_test_ext() -> sp_io::TestExternalities {
-	frame_system::GenesisConfig::default()
-		.build_storage::<Test>()
-		.unwrap()
-		.into()
-}
+use crate::{self as ctype, mock::*};
+use did::mock as did_mock;
 
 #[test]
 fn it_works_for_default_value() {
-	new_test_ext().execute_with(|| {
-		let pair = ed25519::Pair::from_seed(&*b"Alice                           ");
-		let ctype_hash = H256::from_low_u64_be(1);
-		let account = MultiSigner::from(pair.public()).into_account();
-		assert_ok!(CType::add(Origin::signed(account.clone()), ctype_hash));
-		assert_eq!(<CTYPEs<Test>>::contains_key(ctype_hash), true);
-		assert_eq!(CType::ctypes(ctype_hash), Some(account.clone()));
-		assert_noop!(
-			CType::add(Origin::signed(account), ctype_hash),
-			Error::<Test>::AlreadyExists
-		);
-	});
+	let did_auth_key = did_mock::get_ed25519_authentication_key(true);
+	let did_enc_key = did_mock::get_ed25519_authentication_key(true);
+	let did_att_key = did_mock::get_ed25519_authentication_key(true);
 }
+
+// new_test_ext().execute_with(|| {
+// 	let pair = ed25519::Pair::from_seed(&*b"Alice                           ");
+// 	let ctype_hash = H256::from_low_u64_be(1);
+// 	let account = MultiSigner::from(pair.public()).into_account();
+// 	assert_ok!(CType::add(Origin::signed(account.clone()), ctype_hash));
+// 	assert_eq!(<CTYPEs<Test>>::contains_key(ctype_hash), true);
+// 	assert_eq!(CType::ctypes(ctype_hash), Some(account.clone()));
+// 	assert_noop!(
+// 		CType::add(Origin::signed(account), ctype_hash),
+// 		Error::<Test>::AlreadyExists
+// 	);
+// });
