+++ pallets/did/src/custom_types_stable.json
@@ -1,0 +1,91 @@
+{
+	"UrlEncoding": "Text",
+	"DidIdentifier": "AccountId",
+	"DidVerificationKeyType": {
+	  "_enum": [
+		"Authentication",
+		"CapabilityDelegation",
+		"CapabilityInvocation",
+		"AssertionMethod"
+	  ]
+	},
+	"DidEncryptionKeyType": {
+	  "_enum": [
+		"KeyAgreement"
+	  ]
+	},
+	"PublicVerificationKey": {
+	  "_enum": {
+		"Ed25519": "[u8; 32]",
+		"Sr25519": "[u8; 32]"
+	  }
+	},
+	"DidSignature": {
+	  "_enum": {
+		"Ed25519": "Ed25519Signature",
+		"Sr25519": "Sr25519Signature"
+	  }
+	},
+	"PublicEncryptionKey": {
+	  "_enum": {
+		"X55519": "[u8; 32]"
+	  }
+	},
+	"DidError": {
+	  "_enum": {
+		"StorageError": "StorageError",
+		"SignatureError": "SignatureError",
+		"OperationError": "OperationError"
+	  }
+	},
+	"StorageError": {
+	  "_enum": {
+		"DidAlreadyPresent": "null",
+		"DidNotPresent": "null",
+		"DidKeyNotPresent": "DidVerificationKeyType",
+		"VerificationKeysNotPresent": "Vec<PublicVerificationKey>"
+	  }
+	},
+	"SignatureError": {
+	  "_enum": [
+		"InvalidSignatureFormat",
+		"InvalidSignature"
+	  ]
+	},
+	"OperationError": {
+	  "_enum": [
+		"InvalidNonce"
+	  ]
+	},
+	"DidCreationOperation": {
+	  "did": "DidIdentifier",
+	  "new_auth_key": "PublicVerificationKey",
+	  "new_key_agreement_key": "PublicEncryptionKey",
+	  "new_attestation_key": "Option<PublicVerificationKey>",
+	  "new_delegation_key": "Option<PublicVerificationKey>",
+	  "new_endpoint_url": "Option<UrlEncoding>"
+	},
+	"DidUpdateOperation": {
+	  "did": "DidIdentifier",
+	  "new_auth_key": "Option<PublicVerificationKey>",
+	  "new_key_agreement_key": "Option<PublicEncryptionKey>",
+	  "new_attestation_key": "Option<PublicVerificationKey>",
+	  "new_delegation_key": "Option<PublicVerificationKey>",
+	  "verification_keys_to_remove": "Option<BTreeSet<PublicVerificationKey>>",
+	  "new_endpoint_url": "Option<UrlEncoding>",
+	  "tx_counter": "u64"
+	},
+	"DidDeletionOperation": {
+	  "did": "DidIdentifier",
+	  "tx_counter": "u64"
+	},
+	"DidDetails": {
+	  "auth_key": "PublicVerificationKey",
+	  "key_agreement_key": "PublicEncryptionKey",
+	  "delegation_key": "Option<PublicVerificationKey>",
+	  "attestation_key": "Option<PublicVerificationKey>",
+	  "verification_keys": "BTreeSet<PublicVerificationKey>",
+	  "endpoint_url": "Option<UrlEncoding>",
+	  "last_tx_counter": "u64"
+	}
+  }
